maps working

+++ company-account-service/app/api/main.py
@@ -2,6 +2,7 @@ from fastapi import FastAPI
 from app.api.configs.configs import Config, environmentSettings
 from starlette.middleware.cors import CORSMiddleware
 import asyncio
+import os
 app = FastAPI(    
     title=Config.application_name
 )
@@ -42,19 +43,15 @@ async_session = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession
 
 @app.on_event("startup")
 async def startup():
-    await asyncio.sleep(5)
-    # metadata.create_all(engine)
-    ...
-    # from app.api.sqlalchemy_models.db import database
-    async with engine.begin() as conn:
-        await conn.run_sync(metadata.drop_all)
-        await conn.run_sync(metadata.create_all)
+    if environmentSettings.ENV == "DEV":
+        if os.getenv('FIRST_START') == 'true' or os.getenv('FIRST_START') is None:
+            import pipes
+            print("export FIRST_START=%s" % (pipes.quote('false')))
+            await asyncio.sleep(10)
+            async with engine.begin() as conn:
+                await conn.run_sync(metadata.drop_all)
+                await conn.run_sync(metadata.create_all)
+            from app.api.routes.test_routes import get_create_user
+            await get_create_user()
 
-    # await database.connect()
-    # from app.api.sqlalchemy_models.db import engine
-    # if environmentSettings.ENV == 'DEV':
-    #     print('setup')
-    #     async with engine.begin() as conn:
-    #         # await conn.run_sync(Base.metadata.drop_all)
-    #         await conn.run_sync(Base.metadata.create_all)
-    #         print("Created")
+
